+++ BussinessObject/Models/CoinTransaction.cs
@@ -15,7 +15,7 @@ public partial class CoinTransaction
 
     public decimal? MoneyFluctuations { get; set; }
 
-    public int? CoinFluctuations { get; set; }
+    public decimal? CoinFluctuations { get; set; }
 
     public DateTime? Date { get; set; }
